+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.0.0-rc.86 (2020-10-12)
+## v3.0.0-rc.88 (2020-10-12)
 
 #### :house: Internal
 
@@ -20,6 +20,19 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
   * `elNameSelectorGenerator`
   * `elModNameGenerator`
   * `elModSelectorGenerator`
+
+## v3.0.0-rc.87 (2020-10-11)
+
+#### :bug: Bug Fix
+
+* Fixed restoring of a functional state `core/component/watch` `core/component/functional`
+
+## v3.0.0-rc.86 (2020-10-11)
+
+#### :bug: Bug Fix
+
+* Fixed an optimization of lazy watchers `base/i-block/modules/field`
+* Fixed immediate watchers `core/component/watch`
 
 ## v3.0.0-rc.85 (2020-10-09)